+++ fichiers/evenement.json
@@ -33,5 +33,12 @@
         "enquete_liee": "00152ba7-a32e-44e8-8552-96c7bd534326",
         "date_evenement": "2024-12-17",
         "lieu": "kot nathan"
+    },
+    {
+        "id": "173aed24-7d7d-4f29-a0b4-ae1476082763",
+        "nom": "discussion comm",
+        "enquete_liee": "2d357b53-2e16-42f1-a153-d8dcf6e9597d",
+        "date_evenement": "2024-12-17",
+        "lieu": "L101"
     }
 ]
