+++ fichiers/personnes.json
@@ -69,7 +69,7 @@
         "classe": "Employe",
         "sexe": "Homme",
         "date_de_naissance": "2003-09-21",
-        "metier": "'Pas de métier actuellement'",
+        "metier": "Chomeur",
         "mail": "ralphgoffar@gmail.com",
         "interrogatoires": {},
         "matricule": 1,
